+++ R/bme_cv.R
@@ -36,7 +36,7 @@
 #'
 #' @usage bme_cv(ch, cs, zh, a, b,
 #'        model, nugget, sill, range, nsmax = 5,
-#'        nhmax = 5, n = 50, zk_range = range(zh, a, b, -2, 2),
+#'        nhmax = 5, n = 50, zk_range = extended_range(zh, a, b),
 #'        type)
 #'
 #' @param ch A matrix of spatial coordinates for hard data locations
@@ -65,24 +65,24 @@
 #'        the observed data (\code{zh}, \code{a}, and \code{b}). It is advisable
 #'        to explore the posterior distribution at a few locations using
 #'        \code{prob_zk()} before finalizing this range. The default is
-#'        \code{c(min(zh, a, -2), max(zh, b, 2))}.
+#'        \code{extended_range(zh, a, b)}.
 #' @param type A string indicating the type of BME prediction to compute: either
 #'        \code{"mean"} for the posterior mean or \code{"mode"} for the
 #'        posterior mode.
 #'
 #' @examples
 #' data("utsnowload")
-#' ch <- data.matrix(utsnowload[2:10, c("latitude", "longitude")])
-#' cs <- data.matrix(utsnowload[68:232, c("latitude", "longitude")])
-#' zh <- c(utsnowload[2:10, c("hard")])
-#' a <- c(utsnowload[68:232, c("lower")])
-#' b <- c(utsnowload[68:232, c("upper")])
+#' ch <- utsnowload[2:10, c("latitude", "longitude")]
+#' cs <- utsnowload[68:232, c("latitude", "longitude")]
+#' zh <- utsnowload[2:10, c("hard")]
+#' a <- utsnowload[68:232, c("lower")]
+#' b <- utsnowload[68:232, c("upper")]
 #' bme_cv(ch, cs, zh, a, b, model = "exp", nugget = 0.0953, sill = 0.3639,
 #'        range = 1.0787, type = "mean")
 #'
 #' @export
 bme_cv <- function(ch, cs, zh, a, b, model, nugget, sill, range, nsmax = 5,
-                   nhmax = 5, n = 50, zk_range = range(zh, a, b, -2, 2),
+                   nhmax = 5, n = 50, zk_range = extended_range(zh, a, b),
                    type) {
   type <- match.arg(type, choices = c("mean", "mode"))
   col_idx <- if (type == "mode") 1 else c(2, 3)