+++ R/prob_zk.R
@@ -20,7 +20,7 @@
 #'
 #' @usage prob_zk(x, ch, cs, zh, a, b,
 #'         model, nugget, sill, range, nsmax = 5,
-#'         nhmax = 5, n = 50, zk_range = range(zh, a, b, -2, 2),
+#'         nhmax = 5, n = 50, zk_range = extended_range(zh, a, b),
 #'         plot = FALSE)
 #'
 #' @param x A two-column matrix of spatial coordinates for a single estimation
@@ -51,16 +51,16 @@
 #'        the observed data (\code{zh}, \code{a}, and \code{b}). It is advisable
 #'        to explore the posterior distribution at a few locations using
 #'        \code{prob_zk()} before finalizing this range. The default is
-#'        \code{c(min(zh, a, -2), max(zh, b, 2)}.
+#'        \code{extended_range(zh, a, b)}.
 #' @param n An integer indicating the number of points at which to evaluate the
 #'        posterior density over \code{zk_range}.
 #' @param plot Logical; if \code{TRUE}, plots the posterior density curve.
 #'
 #' @examples
 #' data("utsnowload")
-#' x <- data.matrix(utsnowload[1, c("latitude", "longitude")])
-#' ch <- data.matrix(utsnowload[2:67, c("latitude", "longitude")])
-#' cs <- data.matrix(utsnowload[68:232, c("latitude", "longitude")])
+#' x <- utsnowload[1, c("latitude", "longitude")]
+#' ch <- utsnowload[2:67, c("latitude", "longitude")]
+#' cs <- utsnowload[68:232, c("latitude", "longitude")]
 #' zh <- utsnowload[2:67, "hard"]
 #' a <- utsnowload[68:232, "lower"]
 #' b <- utsnowload[68:232, "upper"]
@@ -69,7 +69,7 @@
 #'
 #' @export
 prob_zk <- function(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax = 5,
-                    nhmax = 5, n = 50, zk_range = range(zh, a, b, -2, 2),
+                    nhmax = 5, n = 50, zk_range = extended_range(zh, a, b),
                     plot = FALSE) {
 
   check_x(x, cs, ch)
@@ -78,7 +78,12 @@ prob_zk <- function(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax = 5,
   check_vectors(zh, a, b)
   check_lengths(ch, zh, cs, a, b)
 
-  x <- matrix(x, ncol = 2)
+  x  <- clean_input(x)
+  ch <- clean_input(ch)
+  cs <- clean_input(cs)
+  zh <- clean_input(zh)
+  a  <- clean_input(a)
+  b  <- clean_input(b)
 
   if (nrow(x) != 1) {
     stop("Can only compute the mapping set for a single location")
@@ -133,7 +138,7 @@ prob_zk <- function(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax = 5,
 
   # covariance matrix
   cov_a <- cov_s_s - cov_s_h %*% inv_cov_hs_hs %*% cov_h_s
-  if (det(cov_a) <= 0) cov_a <- cov_s_s
+  #if (det(cov_a) <= 0) cov_a <- cov_s_s
 
   # mean vector
   mu_a <- c(cov_s_h %*% inv_cov_hs_hs %*% zh)
@@ -154,7 +159,7 @@ prob_zk <- function(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax = 5,
   # conditional variance
   inv_cov_kh_kh <- solve(cov_kh_kh)
   cov_soft <- cov_s_s - cov_s_kh %*% inv_cov_kh_kh %*% cov_kh_s
-  if (det(cov_soft) <= 0) cov_soft <- cov_s_s
+  #if (det(cov_soft) <= 0) cov_soft <- cov_s_s
 
   for (i in 1:n) {
 
@@ -189,8 +194,8 @@ prob_zk <- function(x, ch, cs, zh, a, b, model, nugget, sill, range, nsmax = 5,
       sigma = cov_soft
     )[1]
 
-    if (f_soft == 0) f_soft <- 1e-4
-    if (aa == 0) aa <- 1e-4
+    #if (f_soft == 0) f_soft <- 1e-4
+    #if (aa == 0) aa <- 1e-4
 
     pk[i] <- round(((1 / aa) * f_zk * f_soft), 5)
   }