+++ tests/testthat/test-bme_predict.R
@@ -2,12 +2,12 @@
 
 # data
 data("utsnowload")
-x <- data.matrix(utsnowload[1, c("latitude", "longitude")])
-ch <- data.matrix(utsnowload[2:67, c("latitude", "longitude")])
-cs <- data.matrix(utsnowload[68:232, c("latitude", "longitude")])
-zh <- c(utsnowload[2:67, c("hard")])
-a <- c(utsnowload[68:232, c("lower")])
-b <- c(utsnowload[68:232, c("upper")])
+x <- utsnowload[1, c("latitude", "longitude")]
+ch <- utsnowload[2:67, c("latitude", "longitude")]
+cs <- utsnowload[68:232, c("latitude", "longitude")]
+zh <- utsnowload[2:67, c("hard")]
+a <- utsnowload[68:232, c("lower")]
+b <- utsnowload[68:232, c("upper")]
 
 
 # test for posterior mode
@@ -19,7 +19,7 @@ test_that("posterior mode function works", {
   k2 <- bme_predict(x, ch, cs, zh, a, b, model = "exp", nugget = 0.0953,
                     sill = 0.3639, range = 1.0787, type = "mode")[[3]]
 
-  expect_equal(round(k1,3), round(k2,3))
+  expect_equal(round(k1,2), round(k2,2))
 })
 
 
@@ -32,5 +32,5 @@ test_that("posterior mean function works", {
   k2 <- bme_predict(x, ch, cs, zh, a, b, model = "exp", nugget = 0.0953,
                     sill = 0.3639, range = 1.0787, type = "mean")[[3]]
 
-  expect_equal(round(k1,3), round(k2,3))
+  expect_equal(round(k1,2), round(k2,2))
 })