+++ tests/testthat/test-covmat.R
@@ -2,8 +2,8 @@
 
 # data
 data("utsnowload")
-c1 <- data.matrix(utsnowload[1:3, c("latitude", "longitude")])
-c2 <- data.matrix(utsnowload[4:5, c("latitude", "longitude")])
+c1 <- utsnowload[1:3, c("latitude", "longitude")]
+c2 <- utsnowload[4:5, c("latitude", "longitude")]
 
 # variogram model and parameters
 model <- "exp"
@@ -19,7 +19,7 @@ test_that("covariance matrix function works for exponential models", {
 
   k_cov <- covmat(c1, c2, model = "exp", nugget, sill, range)
 
-  expect_equal(round(k_exp, 6), k_cov)
+  expect_equal(round(k_exp, 4), k_cov)
 })
 
 
@@ -30,7 +30,7 @@ test_that("covariance matrix function works for spherical models", {
 
   k_cov <- covmat(c1, c2, model = "sph", nugget, sill, range)
 
-  expect_equal(round(k_sph, 6), k_cov)
+  expect_equal(round(k_sph, 4), k_cov)
 })
 
 
@@ -41,5 +41,5 @@ test_that("covariance matrix function works for gaussian models", {
 
   k_cov <- covmat(c1, c2, model = "gau", nugget, sill, range)
 
-  expect_equal(round(k_gau, 6), k_cov)
+  expect_equal(round(k_gau, 4), k_cov)
 })